requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
packaging==24.0
pluggy==1.5.0
PyMySQL==1.1.1
pytest==8.2.2
pytest-mock==3.14.0
python-dotenv==1.0.1
Werkzeug==3.0.3
